Pull Request Tryout (Insert into empInfo SQL) (#3)
Here is the SQL function that will add info to just the empInfo table. This is me assuming that a button event will call for each string method to run through a prepared statement and insert values straight from the JTextfields. If this format is fine then I'll add the rest of the functions this way or we can do this with objects like your examples . Also let me know if the way I posted this change was alright